+++ Neat/src/Scene/EditorCamera.hpp
@@ -58,7 +58,7 @@ class EditorCamera {
 
   float m_distance = 10.0f;
 
-  float m_viewportWidth = 1280;
-  float m_viewportHeight = 720;
+  UInt32 m_viewportWidth = 1280;
+  UInt32 m_viewportHeight = 720;
 };
 }  // namespace Neat
